requirements.txt
flask==3.0.0
google-api-python-client==2.108.0
openai==1.58.1
python-dotenv==1.0.0
httpx==0.27.2
